kgit-checkpoint: delete .gitignore and .mailmap on orphan branch
There was a separate block of code for deleting .gitignore since
it wasn't swept up in the output of "ls" in the above block.
However this was failing to remove anything, and was only noticed
when adding additional error checking, i.e:

Switched to a new branch 'meta-ovp'
fatal: pathspec 'meta-ovp/.gitignore' did not match any files

This happens because .gitignore isn't in the index of the orphan.
We can just delete the git version of the removal block and add
it to the output of ls (along with .mailmap -- we don't want that
on the meta branch either.)

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>