kgit-publish: use mktemp for intermediary tree
We currently don't clean up from a failed tree "publish", and
so the dir of $src.tmp is not unique, and may exist with some
content.  If it exists, the attempt to re-use it can fail, as
shown below:

  treegen$KMETA=meta-ovp kgit-publish -c -a tree tree-pub
  Switched to branch 'meta-ovp'
  [meta-ovp b0e9067] meta: add README
   1 file changed, 1 insertion(+)
  [ERROR]: git commit 00-README failed
  [ERROR] checkpoint failed

The fail is because $src.tmp already exists and has some/all of
the intended commits.  Rather than nuke-and-pave over the old
dir, simply create a unique one, since the old data may be key
in debugging random issues.

Also add error checking on the copy command.

Signed-off-by: Paul Gortmaker <paul.gortmaker@windriver.com>