updateme: improve duplicate patch detection and performance
We only need to migrate a directory (and its subdirs) once. To improve
performance, we can check to see if a patch is present in the destination
directory structure, and if present, there's no need to do a migration.

Similarly, we can avoid adding duplicate patches, but still have them
processed, by checking for a patch being in a .scc file and only generating
if it isn't already described by a feature.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>