buildall: drop NOMETAGEN
The following commit removes the meta-series files that are required
for skipping the meta gen step, so that NOMETAGEN is no longer valid.
"It's dead code. The meta-series are gone, and will stay gone from the
tree":

commit b08749d4345506d442ac5196b20c003121c6aee6, "kgit-scc: remove -meta
files after consruction"

This is a partial revert of commit 02be3b5f102357a0c0b86f0861f0f16a9d6e5299,
 "buildall: switch back to scc driven processing"

Signed-off-by: Michel Thebeau <michel.thebeau@windriver.com>