updateme: anchor target search regex
The regex that confirms that a found .scc target matches the target machine
is more specific than the regex that searches the .scc files. As a result,
we can match on the wrong target, and fail the second level check, which
results in an autogenerated (and likely bad) .scc file.

By anchoring the search regex with whitespace and the end of line, we'll
no longer mistake: "target-foo", when we are really looking for "target".

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>