tools: allow meta directories that are not the same as the branch name
With this change it is now possible to have a meta branch with meta data
in a directory that is not the same name as the branch.

The changes to three parts of the build are required to discover the name
of the meta directory by relying on the fact that in a clean/proper build
the meta directory is the only untracked, top level directory in the build.

As such, we can restore a checkpoint and then examine the build directory
to determine the meta directory name .. avoiding any new variables to
indicate this to the scripts and build system.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>