kgit-s2q: move subject and diffstat mismatch to 'fuzzy' matching
When a patch on the top of a branch shares a shortlog with a patch
further down in the queue, the default behaviour is to do a secondary
compare on the diffstat. If they don't match, we warn, but pick the
patch as the resume point anyway.

This behaviour is triggering some false positives, and hence patch
failures when the wrong fence post is picked.

We switch the default to only warn on this type of situation, and
if a resume really is intended the --fuzzy flag can be passed to
restore the old behaviour.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>