kgit: retain trailing / in directory processing
In order to limit the matching of path relocation and sanity checking to
actual directory divisions, we need to retain the trailing / on paths. We
really only care about // and other invalid characters.

Once we are sure the trailing / is present, we'll no longer match on
partial directory names when multiple directories contain a common
payth and final directory name.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>