createme: determine meta directory by checking untracked dirs
The previous technique of checking for the meta directory, meant that
only meta directories of the specific .meta* format could be found, and
if no dirs were present at all, a concerning warning was generated.

By syncing createme to use the same technique as the other scripts, and
using git to return untracked directories, we can locate the meta directory
no matter what name it takes.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>