createme: remove meta branch checks
Before the kernel's meta data became an orphan branch, there was a mode
where the meta data could be merged into all branches. To determine if
meta data needed to be checked out, or not, a test was added to see if
the meta data branch was contained by all others.

This check adds 5 seconds each and every build. Since this mode is no
longer supported, we can save the time and delete the check.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>