kgit-scc: use checkpointed meta directory (if available) during treegen
When a history complete treegen is being performed, the meta data respository
is imported and its history grafted onto the kernel tree. When we operate in
this mode, the tree is then checkpoint restored before the generation of the
tree is performed.

A checkpointed restored tree has meta data in .$META, not $META. So we switch
to using the dot directory when this is detected.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>