kgit-meta: always clear series file on branch transitions
When changing branches it used to be a purely addititive operation
if a branch existed. This is no longer the case, so we shouldn't
preserve the existig series file, but should instead clear the series
file, and let the patch routines add to it if required.

This was triggered by the patch optimization changes, that no longer
run do_patch if a leaf/final branch is not being processed.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>