kconf_check: make mismatch details accessible
After a mismatch (a kernel configuration option that doesn't make the
final .config) is found, reporting tools may need access to the details
without needing to mine down into the configuration audit directory
structure.

To faciliate this, we make a copy of mismatch.cfg (and any other files
that we deem reportable) to the top level meta directory (along side
the meta-series).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>