kgit-meta: skip patches on non-leaf nodes
In a similar way as commit 0768d697 [kgit-meta: dont run kgit-s2q for
non-leaf nodes], we can save even more processing time by not even
analysing and linking patches if we aren't on the leaf node of the
tree.

This early exit can save nearly 95% of the time required to "patch"
a tree when no changes are actually applied.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>