configme: standalone operation
configme relies on kgit-meta to gather fragments for processing.
To allow configuration of an external meta-data repository and non
build system use cases, we need a single script with no dependencies
or complex command lines.

By moving the fragment gathering routines into configme, we have the
first step in this process complte.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>