patchme/updateme: unify meta directory handling
With the split of kernel meta from the main repo, and the ability to
run the tools in various different orders, we need to ensure that all
the scripts handle the detectiona and setting of the meta directory
the same way.

With this, we've sync'd the final two scripts to that way of handling
meta dirs.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>