updateme: avoid double inclusion of .scc files
When a .cfg is passed via the command line, the tools check to see if
a .scc file already include that fragment, and then avoid generating
a feature for that fragment (since it already has an explicit .scc
structure).

The tools then ensure that the .scc file is put into the meta-series
processing pipeline .. so the configuration fragment will be applied.

Unfortunately, this forced inclusion of the .scc file in the pipeline
can lead to duplicate processing, since there are other includes and
structure of the files in play.

To avoid the duplicate processing, if a .cfg file is already included
by a fragment, we trust the structure of the includes and don't force
it to be processed.

[YOCTO: #8486]

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>