kgit: add --meta option
The code to find a meta directory was spread across many different
scripts. This obviously is a maintenance issue.

We centralize that code in the main kgit script and allow other
scripts to either source kgit and use the function, or call kgit --meta
from the command line.

Along with this new routine, we update the help message and add command
line processing to the script (since we can how pass options).

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>