kconf_check: standalone operation and command line options
kconf_check was coupled to the output of the other kern tools, in this
commit we break that coupling and allow fully standalone operation of
the script.

As such, kconf_check now only needs the following inputs:

 - a linux kernel source tree
 - a linux processed .config file
 - a list of fragments or command line options

With these inputs a report is generated that finds common errors:

 - redefined options
 - dropped options
 - invalid options
 - repeated options (same fragment)

If the linux source directory includes meta data it will be detected
and used to classify the configuration options. This allows the reports
to be targetted to flagged options (versus reporting them all), but
meta data is completely optional and transparent to the end user.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>