kgit-meta: adapt to new scc/spp meta series generation
Now that scc/spp and the other tools are modular (and smarter),
kgit-meta can be drastically simplified. It no longer needs to
worry about git, branches, kconfigs, or anything but taking
a meta series and applying it to a tree.

Currently this means branches and patches.

Additionally, kgit-meta is only used to generate a tree, so we
don't have to worry about it resuming or incorrect apply patches
when a developer is doing a build via a build system.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>