tools: import Kconfiglib
To provide better analysis of config items that don't make it into
the final .config, we can use Kconfiglib to load the .config and
parse the kernel's Kconfig files.

To avoid introducing a library dependency, we grab a copy of
Kconfiglib here and then make it available where ever the kern-tools
can be used.

Kconfiglib works standalone, so the capture makes sense, versus
looking for build system, or distro specific packaging.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>