kconf_check: use symbol_why to provide diagnostics on missing options
When a symbol is missing from the final .config, kconf_check can warn
the user (and point to the fragment that specified the option), but
figuring out why that option was dropped is left up to the end user.

This process normally involves going to the kernel source tree,
grepping for the config option and looking at the dependencies
(either in the source, or via menuconfig).

With Kconfiglib, we can load the Kconfig structure from the tree,
apply the active .config and provide diagnostics about the Kconfig
item, its dependencies and values of those dependencies. This allows
the user to fix most problems, without going to the source and
manually checking.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>