s2q: always log the current patch
s2q knows how to resume by looking at file in .git that logs the
last applied patch.

In situtations where a patch fails to apply (and is then resolved)
the series application won't resume properly since the last
applied patch was never logged.

We remove all the conditional logging of applied patches and simply
log each patch before it is applied. This ensures that we always
have the right thing tracked.

Signed-off-by: Bruce Ashfield <bruce.ashfield@windriver.com>